requirements.txt
azure_storage==0.37.0
pandas==1.5.3
azure-storage-blob